Data.BoundedSize.Class

Description

Use this module when you need to add an IsBoundedSize instance to a type.

Documentation

class HasSize a where

Size is an abstract Int-valued property of a Type that can be used to refine values of that type.

Methods

size :: a -> Int

class (KnownNat l, KnownNat u, HasSize a) => IsBoundedSize l u a

Class of types which can be assigned a type-level minimum and maximum size.

Associated Types

data BoundedSize l u a

class IsBoundedSize l l a => IsFixedSize l a

Class of types which can be assigned a fixed type-level size.

type FixedSize n a = BoundedSize n n a

class IsBoundedSize 0 l a => IsMaxSize l a

Class of types which can be assigned a maximum type-level size.

type MaxSize n a = BoundedSize 0 n a

type AtLeast c m = forall l. (KnownNat l, m <= l) => c l

Sometimes a MaxSize value is given, but the only thing known about the "max" is that the value is legal, e.g. "foo" could be a MaxSize 3 String, or a MaxSize 4 String, so it is AtLeast MaxSize 3 Then it can be later cast to a specific MaxSize (of at least 3).