Data.ByteString.Integral

Description

Conversions between numbers and bytestrings used to represent them.

3.2.1.3 Messaging Data Types For stream encoding, because of the manner in which data is encoded and transferred across the interface, the types used in method parameter and result values are described using two basic types: a. Byte-string values are a sequence of n bytes that are used to represent strings, blobs, bit vectors, etc. b. N length integer values are whole numbers that are either signed or unsigned.

rollUp :: Bits a => (Word8 -> a) -> (Word8, ByteString) -> a

byteStringToNatural :: ByteString -> Natural

byteStringToInteger :: ByteString -> Integer

naturalToByteString :: Natural -> ByteString

integerToByteString :: Integer -> ByteString

word8 :: Integral a => a -> Word8

char :: Word8 -> Char

ordw :: Char -> Word8

byte :: Integral a => a -> Word8

int :: Integral a => a -> Int

integer :: Integral a => a -> Integer

natural :: Integral a => a -> Natural

class IsByteString a where

Methods

fromByteString :: ByteString -> a

toByteString :: a -> ByteString