Data.HasSize

Description

Things that have a size

Documentation

class HasSize a where

size is an abstract Int-valued property of a type that can be used to refine values of that type.

The size of an Integer is roughly the number of bytes required to represent

The size of a ListLike is just the number of Items

Methods

size :: a -> Int