Data.SizedStaticText

Description

sized-text combinators are defined for members of IsSizedStaticText class. The package includes IsSizedStaticText instances for several common types.

This module is meant to be imported qualifed, e.g.

import qualified Data.SizedStaticText as S

Constructing sized texts

See also unsafeCreate

createLeft :: forall a l u. (IsSizedStaticText a, KnownNat l, KnownNat u, l <= u) => Elem a -> a -> SizedStatic a l u

Elements on the left are preferred. >>> createLeft ' ' "foobarbaz" :: SizedStatic String 0 6 "foobar" >>> createLeft '@' "foobarbaz" :: SizedStatic String 12 20 "foobarbaz@@@"

createRight :: forall a l u. (IsSizedStaticText a, KnownNat l, KnownNat u, l <= u) => Elem a -> a -> SizedStatic a l u

Just like createLeft, except that elements on the right are preferred. >>> createRight '@' "foobarbaz" :: SizedStatic String 0 6 "barbaz" >>> createRight # "foobarbaz" :: SizedStatic String 12 20 "###foobarbaz"

sz :: LitS -> Q Exp

Type-safe SizedStatic constructor macro for string literals.

Example:

$(sz "Foobar")

compiles to

unsafeCreate "Foobar" :: forall a. (IsString a, IsSizedStaticText a) => SizedStatic a 6 6

where 6 is the string length obtained at compile time.

create :: forall a (l :: Nat) (u :: Nat). (IsSizedStaticText a, KnownNat l, KnownNat u, l <= u) => a -> Maybe (SizedStatic a l u)

Attempt to safely create a SizedStatic if it matches target length.

>>> create "foobar" :: Maybe (SizedStatic String 6 10)
Just "foobar"
>>> create "barbaz" :: Maybe (SizedStatic String 0 4)
Nothing

This is safer than unsafeCreate and unlike with createLeft createRight the source value is left unchanged. However, this implies a further run-time check for Nothing values.

replicate :: forall a l u. (IsSizedStaticText a, KnownNat l, KnownNat u, l <= u) => Elem a -> SizedStatic a l u

Construct a new SizedStatic of maximum length from a basic element.

>>> replicate '=' :: SizedStatic String 5 10
"=========="

Working with sized texts

append :: forall a l1 u1 l2 u2. (IsSizedStaticText a, KnownNat l1, KnownNat u1, KnownNat l2, KnownNat u2, KnownNat (l1 + l2), KnownNat (u1 + u2)) => SizedStatic a l1 u1 -> SizedStatic a l2 u2 -> SizedStatic a (l1 + l2) (u1 + u2)

Append two SizedStatics together.

>>> append $(sz "foo") $(sz "bar") :: SizedStatic String 6 6
"foobar"
>>> :set -fno-warn-type-defaults
>>> append $(sz "Hello, ") $(sz "world!")
"Hello, world!"
>>> :type it
it
  :: (IsSizedStaticText a, Data.String.IsString a) =>
     SizedStatic a 13 13

take :: forall a l1 u1 l2 u2. (IsSizedStaticText a, KnownNat l1, KnownNat u1, KnownNat l2, KnownNat u2, l2 <= l1) => SizedStatic a l1 u1 -> SizedStatic a l2 u2

Reduce SizedStatic length, preferring elements on the left.

>>> take $(sz "Foobar") :: SizedStatic String 1 3
"Foo"

drop :: forall a l1 u1 l2 u2. (IsSizedStaticText a, KnownNat l1, KnownNat u1, KnownNat l2, KnownNat u2, l2 <= l1) => SizedStatic a l1 u1 -> SizedStatic a l2 u2

Reduce SizedStatic length, preferring elements on the right.

>>> drop $(sz "Foobar") :: SizedStatic String 0 2
"ar"

map :: IsSizedStaticText a => (Elem a -> Elem a) -> SizedStatic a l u -> SizedStatic a l u

Map a SizedStatic to a SizedStatic of the same length.

>>> map toUpper $(sz "Hello") :: SizedStatic String 5 5
"HELLO"

padLeft :: forall a l1 u1 l2 u2. (IsSizedStaticText a, KnownNat l1, KnownNat u1, KnownNat l2, KnownNat u2, KnownNat (u2 + u1), l2 <= l1, u1 <= u2) => Elem a -> SizedStatic a l1 u1 -> SizedStatic a l2 u2

Fill a SizedStatic with extra elements up to target length, padding original elements to the left.

padRight :: forall a l1 u1 l2 u2. (IsSizedStaticText a, KnownNat l1, KnownNat u1, KnownNat l2, KnownNat u2, KnownNat (u1 + u2), l2 <= l1, u1 <= u2) => Elem a -> SizedStatic a l1 u1 -> SizedStatic a l2 u2

Like padLeft, but original elements are padded to the right.

bounds :: forall a l u. (IsSizedStaticText a, KnownNat l, KnownNat u) => SizedStatic a l u -> (Int, Int)

Obtain length bounds from the type.

length :: forall a l u. (IsSizedStaticText a, KnownNat l, KnownNat u) => SizedStatic a l u -> Int

Obtain value-level length. Consult the actual data value.

>>> length $ (take $(sz "password") :: SizedStatic String 4 32)
8

StaticText conversions

fromStatic :: (IsStaticText a, IsSizedStaticText a) => Static a i -> SizedStatic a i i

toStatic :: (IsSizedStaticText a, IsStaticText a) => SizedStatic a i i -> Static a i

IsSizedStaticText class

data family SizedStatic a (l :: Nat) (u :: Nat)

Data family which wraps values of the underlying type giving them type-level length bounds. SizedStatic t 6 10 means a value of type t of length between 6 and 10.

class IsSizedStaticText a where

Class of types which can be assigned a type-level minimum and maximum length.

Minimal complete definition

unsafeCreate, unwrap, length, append, replicate, map, take, drop

Associated Types

type Elem a

Basic element type. For IsSizedStaticText [a], this is a.

Methods

unsafeCreate :: a -> SizedStatic a l u

Simply wrap a value in a SizedStatic as is, assuming any length.

WARNING Use it only when you know what you're doing.

For example, an expression like

unsafeCreate "somestring" :: SizedStatic String 50 100

will typecheck, although the stored length information will not match actual string size. This may result in wrong behaviour of all functions defined for IsSizedStaticText.

When writing new IsSizedStaticText instances, make this simply apply the constructor of SizedStatic.

unwrap :: SizedStatic a l u -> a

Obtain the underlying value.