Data.SizedText

Description

sized-text combinators are defined for members of IsSizedText class. The package includes IsSizedText instances for several common types.

This module is meant to be imported qualifed, e.g.

import qualified Data.SizedText as S

Constructing sized texts

See also unsafeCreate

createLeft :: forall a l u. (IsSizedText a, KnownNat l, KnownNat u) => Elem a -> a -> Sized a l u

Elements on the left are preferred. >>> createLeft ' ' "foobarbaz" :: Sized String 0 6 "foobar" >>> createLeft '@' "foobarbaz" :: Sized String 12 20 "foobarbaz@@@"

createRight :: forall a l u. (IsSizedText a, KnownNat l, KnownNat u) => Elem a -> a -> Sized a l u

Just like createLeft, except that elements on the right are preferred. >>> createRight '@' "foobarbaz" :: Sized String 0 6 "barbaz" >>> createRight # "foobarbaz" :: Sized String 12 20 "###foobarbaz"

sz :: LitS -> Q Exp

Type-safe constructor for bounded-length string literals: Sized a 0 l

>>> $(sz "Foobar")
"Foobar"

>>> szFoo = $(sz "Foobar")
>>> :type szFoo
szFoo :: (IsString a, IsSizedText a) => Sized a 0 6

>>> runQ $ ppr <$> sz "Foobar"
Data.SizedText.Class.unsafeCreate "Foobar" :: forall a_0 . (Data.String.IsString a_0,
                                                            Data.SizedText.Class.IsSizedText a_0) =>
                                              Data.SizedText.Class.Sized a_0 0 6

>>> :t sz "Foobar"
sz "Foobar" :: Q Exp

where 6 is the string length obtained at compile time.

create :: forall a (l :: Nat) (u :: Nat). (IsSizedText a, KnownNat l, KnownNat u) => a -> Maybe (Sized a l u)

Attempt to safely create a Sized if it matches target length.

>>> create "foobar" :: Maybe (Sized String 6 10)
Just "foobar"
>>> create "barbaz" :: Maybe (Sized String 0 4)
Nothing

This is safer than unsafeCreate and unlike with createLeft createRight the source value is left unchanged. However, this implies a further run-time check for Nothing values.

replicate :: forall a l u. (IsSizedText a, KnownNat l, KnownNat u) => Elem a -> Sized a l u

Construct a new C.Sized of maximum length from a basic element.

>>> replicate '=' :: C.Sized String 5 10
"=========="

Working with sized texts

append :: forall a l1 u1 l2 u2. (IsSizedText a, KnownNat l1, KnownNat u1, KnownNat l2, KnownNat u2) => Sized a l1 u1 -> Sized a l2 u2 -> Sized a (l1 + l2) (u1 + u2)

Append two C.Sizeds together.

>>> :type append $(sz "foo") $(st "bear")
append $(sz "foo") $(st "bear")
  :: (C.IsSizedText a, IsString a) => C.Sized a 4 7
>>> append $(sz "foo") $(st "bear")
"foobear"
>>> :type append $(st "Hello, ") $(sz "world!")
append $(st "Hello, ") $(sz "world!")
  :: (C.IsSizedText a, IsString a) => C.Sized a 7 13
>>> append $(st "Hello, ") $(sz "world!")
"Hello, world!"

take :: forall a l1 u1 l2 u2. (IsSizedText a, KnownNat l1, KnownNat u1, KnownNat l2, KnownNat u2, l2 <= l1) => Sized a l1 u1 -> Sized a l2 u2

Reduce C.Sized length, preferring elements on the left.

>>> take $(sz "Foobar") :: C.Sized String 0 3
"Foo"

drop :: forall a l1 u1 l2 u2. (IsSizedText a, KnownNat l1, KnownNat u1, KnownNat l2, KnownNat u2, l2 <= l1) => Sized a l1 u1 -> Sized a l2 u2

Reduce C.Sized length, preferring elements on the right.

>>> drop $(st "Foobar") :: C.Sized String 2 2
"ar"
>>> drop $(sz "Foobar") :: C.Sized String 0 2
"ar"

map :: IsSizedText a => (Elem a -> Elem a) -> Sized a l u -> Sized a l u

Map a C.Sized to a C.Sized of the same length.

>>> map toUpper $(st "Hello") :: C.Sized String 5 5
"HELLO"

padLeft :: forall a l1 u1 l2 u2. (IsSizedText a, KnownNat l1, KnownNat u1, KnownNat l2, KnownNat u2, KnownNat (u2 + u1), l2 <= l1, u1 <= u2) => Elem a -> Sized a l1 u1 -> Sized a l2 u2

Fill a C.Sized with extra elements up to target length, padding original elements to the left.

padRight :: forall a l1 u1 l2 u2. (IsSizedText a, KnownNat l1, KnownNat u1, KnownNat l2, KnownNat u2, KnownNat (u1 + u2), l2 <= l1, u1 <= u2) => Elem a -> Sized a l1 u1 -> Sized a l2 u2

Like padLeft, but original elements are padded to the right.

bounds :: forall a l u. (IsSizedText a, KnownNat l, KnownNat u) => Sized a l u -> (Int, Int)

Obtain length bounds from the type.

length :: forall a l u. (IsSizedText a, KnownNat l, KnownNat u) => Sized a l u -> Int

Obtain value-level length. Consult the actual data value.

C.IsSizedText class

data family Sized a (l :: Nat) (u :: Nat)

Data family which wraps values of the underlying type giving them a type-level length. Sized t 6 10 means a value of type t of length between 6 and 10.

class IsSizedText a where

Class of types which can be assigned a type-level minimum and maximum length.

Minimal complete definition

unsafeCreate, unwrap, length, append, replicate, map, take, drop

Associated Types

type Elem a

Basic element type. For IsSizedText [a], this is a.

Methods

unsafeCreate :: a -> Sized a l u

Simply wrap a value in a Sized as is, assuming any length.

WARNING Use it only when you know what you're doing.

For example, an expression like

unsafeCreate "somestring" :: Sized String 50 100

will typecheck, although the stored length information will not match actual string size. This may result in wrong behaviour of all functions defined for IsSizedText.

When writing new IsSizedText instances, make this simply apply the constructor of Sized.

unwrap :: Sized a l u -> a

Forget type-level minimum and maximum length, obtaining the underlying value.