Data.SizedText.Class

Description

Use this module when you need to add an IsSizedText instance to a type.

Documentation

class IsSizedText a where

Class of types which can be assigned a type-level minimum and maximum length.

Associated Types

data Sized a (l :: Nat) (u :: Nat)

Data family which wraps values of the underlying type giving them a type-level length. Sized t 6 10 means a value of type t of length between 6 and 10.

type Elem a

Basic element type. For IsSizedText [a], this is a.

Methods

unsafeCreate :: a -> Sized a l u

Simply wrap a value in a Sized as is, assuming any length.

WARNING Use it only when you know what you're doing.

For example, an expression like

unsafeCreate "somestring" :: Sized String 50 100

will typecheck, although the stored length information will not match actual string size. This may result in wrong behaviour of all functions defined for IsSizedText.

When writing new IsSizedText instances, make this simply apply the constructor of Sized.

unwrap :: Sized a l u -> a

Forget type-level length, obtaining the underlying value.

length :: a -> Int

append :: a -> a -> a

replicate :: Int -> Elem a -> a

map :: (Elem a -> Elem a) -> a -> a

take :: Int -> a -> a

drop :: Int -> a -> a