Data.Smart

Description

Smart constructors refined with predicates.

Documentation

class Smart c a where

Smart constructor.

Wrap a a in a c under the control of a refinement predicate.

Minimal complete definition

unwrap, unsafeCreate, predicate

Associated Types

type ErrorMessage c a

Refinement predicate error message type. Commonly would be String.

Methods

unwrap :: c a -> a

Forget type-level minimum and maximum size, obtaining the underlying value.

unsafeCreate :: a -> c a

Simply wrap a value in a c as is, assuming predicate.

WARNING Use it only when you know what you're doing.

predicate :: c a -> Either (ErrorMessage c a) (c a)

Refinement predicate. Either pass a wrapped value as Right or give an error message as Left.

Note that this is passed the value after being wrapped, and therefore assumes that the wrapping operation is mostly harmless.

safeCreate :: a -> Either (ErrorMessage c a) (c a)

Primary smart constructor into Either.

create :: a -> Maybe (c a)

Simplified smart constructor into Maybe.