Extras.Bytes

Description

Datatypes for Tokens.

Documentation

data Fixed_bytes n

Constructors

Fixed_bytes !(Static ShortByteString n)

data family Static a (i :: Nat) :: Type

Data family which wraps values of the underlying type giving them a type-level length. Static t 6 means a value of type t of length 6.

unwrap :: Fixed_bytes n -> ByteString

create :: KnownNat n => ByteString -> Maybe (Fixed_bytes n)

take :: (KnownNat m, KnownNat n, n <= m) => Fixed_bytes m -> Fixed_bytes n

drop :: (KnownNat m, KnownNat n, n <= m) => Fixed_bytes m -> Fixed_bytes n

append :: Fixed_bytes m -> Fixed_bytes n -> Fixed_bytes (m + n)

fixed :: KnownNat n => ByteString -> Fixed_bytes n

fpack :: KnownNat n => [Word8] -> Fixed_bytes n

naturalToFixedBytes :: KnownNat n => Natural -> Fixed_bytes n

Orphan instances

HasHex ShortByteString

Instance details

Methods

hex :: ShortByteString -> String

fromHex :: String -> Maybe ShortByteString

KnownNat n => IsString (Static ShortByteString n)

Instance details

Methods

fromString :: String -> Static ShortByteString n

KnownNat n => HasHex (Static ShortByteString n)

Instance details

Methods

hex :: Static ShortByteString n -> String

fromHex :: String -> Maybe (Static ShortByteString n)