System.SED.Common.TableUIDs

Description

Table UIDs as specified in Table 240.

The specification for each table is translated to three constant declarations. For instance, the information on the AccessControl table becomes

    --| The HalfUID that identifies the AccessControl table
    hAccessControl :: HalfUID
    hAccessControl = halfUID 0x00 0x00 0x00 0x07

    --| The UID that identifies the AccessControl table itself
    uAccessControlTable :: UID
    uAccessControlTable = uid 0x00 0x00 0x00 0x07 0x00 0x00 0x00 0x00

    --| The UID that identifies the AccessControl table entry in the Table Table
    uAccessControlTableObject :: UID
    uAccessControlTableObject = uid 0x00 0x00 0x00 0x01 0x00 0x00 0x00 0x07

Unfortunately, Template Haskell can neither see manual comments nor create comments for Haddock to see; hence the names themselves must suffice as documentation.

Table HalfUID, UIDs, and Object UIDs

For each row of the table (after the header row), define the indicated constants.