System.SED.MCTP.Common.Base_Type

Description

Formats.

Documentation

newtype Core_integer n

Constructors

Core_integer (Integer ? MaxSize n)

newtype Core_uinteger n

Constructors

Core_uinteger (Natural ? MaxSize n)

newtype Core_bytes n

Constructors

Core_bytes (ByteString ? FixedSize n)

newtype Core_max_bytes n

Constructors

Core_max_bytes (ByteString ? MaxSize n)

safeCreate :: Predicate p a => a -> Either String (Refined p a)

toList :: IsList l => l -> [Item l]

The toList function extracts a list of Item l from the structure l. It should satisfy fromList . toList = id.

append :: forall l1 u1 l2 u2 a. (IsBoundedSizeListOps l1 u1 a, IsBoundedSizeListOps l2 u2 a, IsBoundedSizeListOps (l1 + l2) (u1 + u2) a) => (a ? BoundedSize l1 u1) -> (a ? BoundedSize l2 u2) -> a ? BoundedSize (l1 + l2) (u1 + u2)

Append two BoundedSizes together.

>>> append $(mx "foo") $(fx "bear")
"foobear"

>>> :type append $(mx "foo") $(fx "bear")
append $(mx "foo") $(fx "bear")
  :: (Data.ListLike.Base.ListLike a (Item a), HasSize a,
      Data.String.IsString a) =>
     a ? BoundedSize 4 7

>>> append $(fx "Hello, ") $(mx "world!")
"Hello, world!"

replicate :: forall l u a. IsBoundedSizeListOps l u a => Item a -> a ? BoundedSize l u

Construct a new BoundedSize of maximum length from a basic element.

>>> replicate '=' :: String ? BoundedSize 5 10
"=========="

map :: forall l u a. IsBoundedSizeListOps l u a => (Item a -> Item a) -> (a ? BoundedSize l u) -> a ? BoundedSize l u

Map a BoundedSize to a BoundedSize of the same length.

>>> map toUpper $(fx "Hello") :: String ? BoundedSize 5 5
"HELLO"

take :: forall l1 u1 l2 u2 a. (IsBoundedSizeListOps l1 u1 a, IsBoundedSizeListOps l2 u2 a) => (a ? BoundedSize l1 u1) -> a ? BoundedSize l2 u2

Reduce BoundedSize length, preferring elements on the left.

>>> take $(mx "Foobar") :: String ? BoundedSize 0 3
"Foo"
>>> :t take $(fx "Hello")
take $(fx "Hello")
  :: (Data.ListLike.Base.ListLike a (Item a), KnownNat l2,
      KnownNat u2, HasSize a, Data.String.IsString a) =>
     a ? BoundedSize l2 u2

>>> :t take $(fx "Hello") :: ByteString ? BoundedSize 4 32
take $(fx "Hello") :: ByteString ? BoundedSize 4 32
  :: ByteString ? BoundedSize 4 32
>>> take $(fx "Hello") :: ByteString ? BoundedSize 4 32
"Hello"
>>> length (take $(fx "Hello") :: ByteString ? BoundedSize 4 32)
5
>>> take $(fx "Hello") :: ByteString ? BoundedSize 4 32
"Hello"
>>> take $(fx "Hello") :: ByteString ? BoundedSize 4 4
"Hell"
>>> take $(fx "HelloHelloHelloHelloHelloHelloHello") :: ByteString ? BoundedSize 4 32
"HelloHelloHelloHelloHelloHelloHe"

drop :: forall l1 u1 l2 u2 a. (IsBoundedSizeListOps l1 u1 a, IsBoundedSizeListOps l2 u2 a) => (a ? BoundedSize l1 u1) -> a ? BoundedSize l2 u2

Reduce BoundedSize length, preferring elements on the right.

>>> drop $(fx "Foobar") :: String ? BoundedSize 2 2
"ar"
>>> drop $(mx "Foobar") :: String ? BoundedSize 0 2
"ar"

bounds :: forall l u a. IsBoundedSize l u a => (a ? BoundedSize l u) -> (Int, Int)

Obtain length bounds from the type.

length :: forall l u a. IsBoundedSizeListOps l u a => (a ? BoundedSize l u) -> Int

Obtain value-level length. Consult the actual data value.

padLeft :: forall l1 u1 l2 u2 a. (IsBoundedSizeListOps l1 u1 a, IsBoundedSizeListOps l2 u2 a, IsBoundedSizeListOps 0 (u2 + u1) a) => Item a -> (a ? BoundedSize l1 u1) -> a ? BoundedSize l2 u2

Fill a BoundedSize with extra elements up to target length, padding original elements to the left.

padRight :: forall l1 u1 l2 u2 a. (IsBoundedSizeListOps l1 u1 a, IsBoundedSizeListOps l2 u2 a, IsBoundedSizeListOps 0 (u1 + u2) a) => Item a -> (a ? BoundedSize l1 u1) -> a ? BoundedSize l2 u2

Like padLeft, but original elements are padded to the right.

plain :: Predicate p a => Refined p a -> a

unsafeCreate :: Predicate p a => a -> Refined p a

ci :: LitI -> Q Exp

Type-safe constructor for fixed-length byte literals: Core_bytes n

cu :: LitI -> Q Exp

cb :: LitS -> Q Exp

cm :: LitS -> Q Exp