System.SED.MCTP.Common.ColumnTypes.TH

Description

Template Haskell for parsing Table column types in Section 5.1.3.

Documentation

ttype :: QuasiQuoter

QuasiQuoter for Type Table row tables.

Each row of the Type Table is in its own table in the Core Spec.

ttypeDecs :: String -> [Dec]

typeTableParser :: Parser TypeTableRow

typeTableRow :: [Int] -> Parser TypeTableRow

dTypeTableRow :: TypeTableRow -> TypeTableRowDecs

typeTableTitle :: Parser ByteString

spaces :: Parser ByteString

rowSepFieldLengths :: Parser [Int]

rowSep :: Parser [Int]

blankLines :: Parser ()

tableRowFields :: [Int] -> Parser [ByteString]

data TypeTableRow

Constructors

TypeTableRow TypeUIDField TypeName [FormatString]

type TypeUIDField = ByteString

type TypeName = ByteString

type FormatString = ByteString

header :: [Int] -> Parser ()

formatString :: TypeTableRow -> ByteString

newtype TypeTableRowDecs

Constructors

TypeTableRowDecs [Dec]

tenum :: QuasiQuoter

QuasiQuoter for Type Table enum tables.

Some enumerations are explicated by tables following the Type Tables that refer to them.

tenumDecs :: String -> [Dec]

dEnum :: (String, [EnumRow]) -> [Dec]

enumTableParser :: Parser (String, [EnumRow])

data EnumRow

Constructors

EnumRow String [Int]

enumTableTitle :: Parser String

enumTableRow :: [Int] -> Parser EnumRow