System.SED.MCTP.Common.TableUIDs

Description

Table UIDs as specified in Table 240.

The specification for each table is translated to three constant declarations. For instance, the information on the AccessControl table becomes

    --| The HalfUID that identifies the AccessControl table
    hAccessControl :: HalfUID
    hAccessControl = halfUID 0x00 0x00 0x00 0x07

    --| The UID that identifies the AccessControl table itself
    uAccessControlTable :: UID
    uAccessControlTable = uid 0x00 0x00 0x00 0x07 0x00 0x00 0x00 0x00

    --| The UID that identifies the AccessControl table entry in the Table Table
    uAccessControlTableObject :: UID
    uAccessControlTableObject = uid 0x00 0x00 0x00 0x01 0x00 0x00 0x00 0x07

Unfortunately, Template Haskell can neither see manual comments nor create comments for Haddock to see; hence the names themselves must suffice as documentation.

Table HalfUID, UIDs, and Object UIDs

For each row of the table (after the header row), define the indicated constants.

nameUID :: Map UID String

nameHalfUID :: Map Object_Table_HalfUID String

othTable :: Object_Table_HalfUID

uTableTable :: UID

uTableTableObject :: UID

othSPInfo :: Object_Table_HalfUID

uSPInfoTable :: UID

uSPInfoTableObject :: UID

othSPTemplates :: Object_Table_HalfUID

uSPTemplatesTable :: UID

uSPTemplatesTableObject :: UID

othColumn :: Object_Table_HalfUID

uColumnTable :: UID

uColumnTableObject :: UID

othType :: Object_Table_HalfUID

uTypeTable :: UID

uTypeTableObject :: UID

othMethodID :: Object_Table_HalfUID

uMethodIDTable :: UID

uMethodIDTableObject :: UID

othAccessControl :: Object_Table_HalfUID

uAccessControlTable :: UID

uAccessControlTableObject :: UID

othACE :: Object_Table_HalfUID

uACETable :: UID

uACETableObject :: UID

othAuthority :: Object_Table_HalfUID

uAuthorityTable :: UID

uAuthorityTableObject :: UID

othCertificates :: Object_Table_HalfUID

uCertificatesTable :: UID

uCertificatesTableObject :: UID

othC_PIN :: Object_Table_HalfUID

uC_PINTable :: UID

uC_PINTableObject :: UID

othC_RSA_1024 :: Object_Table_HalfUID

uC_RSA_1024Table :: UID

uC_RSA_1024TableObject :: UID

othC_RSA_2048 :: Object_Table_HalfUID

uC_RSA_2048Table :: UID

uC_RSA_2048TableObject :: UID

othC_AES_128 :: Object_Table_HalfUID

uC_AES_128Table :: UID

uC_AES_128TableObject :: UID

othC_AES_256 :: Object_Table_HalfUID

uC_AES_256Table :: UID

uC_AES_256TableObject :: UID

othC_EC_160 :: Object_Table_HalfUID

uC_EC_160Table :: UID

uC_EC_160TableObject :: UID

othC_EC_192 :: Object_Table_HalfUID

uC_EC_192Table :: UID

uC_EC_192TableObject :: UID

othC_EC_224 :: Object_Table_HalfUID

uC_EC_224Table :: UID

uC_EC_224TableObject :: UID

othC_EC_256 :: Object_Table_HalfUID

uC_EC_256Table :: UID

uC_EC_256TableObject :: UID

othC_EC_384 :: Object_Table_HalfUID

uC_EC_384Table :: UID

uC_EC_384TableObject :: UID

othC_EC_521 :: Object_Table_HalfUID

uC_EC_521Table :: UID

uC_EC_521TableObject :: UID

othC_EC_163 :: Object_Table_HalfUID

uC_EC_163Table :: UID

uC_EC_163TableObject :: UID

othC_EC_233 :: Object_Table_HalfUID

uC_EC_233Table :: UID

uC_EC_233TableObject :: UID

othC_EC_283 :: Object_Table_HalfUID

uC_EC_283Table :: UID

uC_EC_283TableObject :: UID

othC_HMAC_160 :: Object_Table_HalfUID

uC_HMAC_160Table :: UID

uC_HMAC_160TableObject :: UID

othC_HMAC_256 :: Object_Table_HalfUID

uC_HMAC_256Table :: UID

uC_HMAC_256TableObject :: UID

othC_HMAC_384 :: Object_Table_HalfUID

uC_HMAC_384Table :: UID

uC_HMAC_384TableObject :: UID

othC_HMAC_512 :: Object_Table_HalfUID

uC_HMAC_512Table :: UID

uC_HMAC_512TableObject :: UID

othSecretProtect :: Object_Table_HalfUID

uSecretProtectTable :: UID

uSecretProtectTableObject :: UID

othTPerInfo :: Object_Table_HalfUID

uTPerInfoTable :: UID

uTPerInfoTableObject :: UID

othCryptoSuite :: Object_Table_HalfUID

uCryptoSuiteTable :: UID

uCryptoSuiteTableObject :: UID

othTemplate :: Object_Table_HalfUID

uTemplateTable :: UID

uTemplateTableObject :: UID

othSP :: Object_Table_HalfUID

uSPTable :: UID

uSPTableObject :: UID

othClockTime :: Object_Table_HalfUID

uClockTimeTable :: UID

uClockTimeTableObject :: UID

othH_SHA_1 :: Object_Table_HalfUID

uH_SHA_1Table :: UID

uH_SHA_1TableObject :: UID

othH_SHA_256 :: Object_Table_HalfUID

uH_SHA_256Table :: UID

uH_SHA_256TableObject :: UID

othH_SHA_384 :: Object_Table_HalfUID

uH_SHA_384Table :: UID

uH_SHA_384TableObject :: UID

othH_SHA_512 :: Object_Table_HalfUID

uH_SHA_512Table :: UID

uH_SHA_512TableObject :: UID

othLog :: Object_Table_HalfUID

uLogTable :: UID

uLogTableObject :: UID

othLogList :: Object_Table_HalfUID

uLogListTable :: UID

uLogListTableObject :: UID

othLockingInfo :: Object_Table_HalfUID

uLockingInfoTable :: UID

uLockingInfoTableObject :: UID

othLocking :: Object_Table_HalfUID

uLockingTable :: UID

uLockingTableObject :: UID

othMBRControl :: Object_Table_HalfUID

uMBRControlTable :: UID

uMBRControlTableObject :: UID

othMBR :: Object_Table_HalfUID

uMBRTable :: UID

uMBRTableObject :: UID

othK_AES_128 :: Object_Table_HalfUID

uK_AES_128Table :: UID

uK_AES_128TableObject :: UID

othK_AES_256 :: Object_Table_HalfUID

uK_AES_256Table :: UID

uK_AES_256TableObject :: UID

Safe accessor functions for the generated Maps

lookupTableHalfUID :: Object_Table_HalfUID -> Maybe String

Look up the name of a Table_Object_HalfUID

lookupTableUID :: UID -> Maybe String

Look up the name of a Table UID or a Table object UID