Data.BoundedSize

Description

sized-text combinators are defined for members of IsBoundedSize class. The package includes IsBoundedSize instances for several common types.

This module is meant to be imported qualifed, e.g.

import qualified Data.BoundedSize as S

Documentation

class IsBytes a where

Class of types with ByteString-like operations

Associated Types

type Elem a

Methods

length :: a -> Int

append :: a -> a -> a

replicate :: Int -> Elem a -> a

map :: (Elem a -> Elem a) -> a -> a

take :: Int -> a -> a

drop :: Int -> a -> a

class IsBoundedSize (l :: Nat) (u :: Nat) a where

Class of types which can be assigned a type-level minimum and maximum length.

Minimal complete definition

unsafeCreate, unwrap, size, predicate

Associated Types

data BoundedSize l u a

Data family which wraps values of the underlying type giving them a type-level size. BoundedSize t 6 10 means a value of type t of size between 6 and 10.

Methods

unsafeCreate :: a -> BoundedSize l u a

Simply wrap a value in a BoundedSize as is, assuming any length.

WARNING Use it only when you know what you're doing.

For example, an expression like

unsafeCreate "somestring" :: BoundedSize String 50 100

will typecheck, although the stored size information will not match actual value size. This may result in wrong behaviour of all functions defined for IsBoundedSize.

When writing new IsBoundedSize instances, make this simply apply the constructor of BoundedSize.

unwrap :: BoundedSize l u a -> a

Forget type-level minimum and maximum size, obtaining the underlying value.

size :: a -> Int

predicate :: (KnownNat l, KnownNat u) => BoundedSize l u a -> Either String (BoundedSize l u a)

safeCreate :: (KnownNat l, KnownNat u) => a -> Either String (BoundedSize l u a)

create :: (KnownNat l, KnownNat u) => a -> Maybe (BoundedSize l u a)

type FixedSize n a = BoundedSize n n a

Class of types which can be assigned a fixed type-level size.

type MaxSize n a = BoundedSize 0 n a

Class of types which can be assigned a maximum type-level size.

class (IsBoundedSize l u a, IsBytes a) => IsBoundedSizeBytes l u a

Class of ByteString-like types which can be assigned a type-level minimum and maximum length.

type BoundedSizeBytes l u a = IsBoundedSizeBytes l u a => BoundedSize l u a

type FixedSizeBytes n a = IsBoundedSizeBytes n n a => FixedSize n a

type MaxSizeBytes n a = IsBoundedSizeBytes n n a => MaxSize n a

type BoundedSizeByteString l u = BoundedSizeBytes l u ByteString

type FixedSizeByteString n = FixedSizeBytes n ByteString

type MaxSizeByteString n = MaxSizeBytes n ByteString

st :: LitS -> Q Exp

Type-safe constructor for fixed-length string literals: BoundedSize a l l

>>> :t "Foobar"
"Foobar" :: IsString p => p
>>> :t LitS "Foobar"
LitS "Foobar" :: LitS
>>> :t st "Foobar"
st "Foobar" :: Q Exp
>>> $(st "Foobar") :: BoundedSize 6 6 ByteString
"Foobar"

>>> runQ $ ppr <$> st "Foobar"
Data.BoundedSize.Class.unsafeCreate "Foobar" :: forall a_0 . (Data.String.IsString a_0,
                                                              Data.BoundedSize.Class.IsBoundedSize 6
                                                                                                   6
                                                                                                   a_0) =>
                                                Data.BoundedSize.Class.BoundedSize 6 6 a_0

>>> :t runQ $ ppr <$> st "Foobar"
runQ $ ppr <$> st "Foobar" :: Quasi m => m Doc

where 6 is the string length obtained at compile time.

sz :: LitS -> Q Exp

Type-safe constructor for bounded-length string literals: BoundedSize a 0 l

>>> :t "Foobar"
"Foobar" :: IsString p => p
>>> :t LitS "Foobar"
LitS "Foobar" :: LitS
>>> :t sz "Foobar"
sz "Foobar" :: Q Exp
>>> $(sz "Foobar") :: BoundedSize 0 6 ByteString
"Foobar"

>>> runQ $ ppr <$> sz "Foobar"
Data.BoundedSize.Class.unsafeCreate "Foobar" :: forall a_0 . (Data.String.IsString a_0,
                                                              Data.BoundedSize.Class.IsBoundedSize 0
                                                                                                   6
                                                                                                   a_0) =>
                                                Data.BoundedSize.Class.BoundedSize 0 6 a_0

>>> :t runQ $ ppr <$> sz "Foobar"
runQ $ ppr <$> sz "Foobar" :: Quasi m => m Doc

where 6 is the string length obtained at compile time.

unsafeCreateExp :: (Type -> Type -> Name -> Type) -> Type -> Type -> Name -> String -> Exp

Construct > unsafeCreate Foobar :: forall a. (IsString a, IsBoundedSize a) => typef a l u where l and u are the type-level KnownNat versions of the bounds of s

>>> at <- runQ $ newName "a"
>>> ppr $ unsafeCreateExp universallyQuantifiedBoundedSizeType (typeFromInt 0) (typeFromInt 4) at "Boo!"
Data.BoundedSize.Class.unsafeCreate "Boo!" :: forall a_0 . (Data.String.IsString a_0,
                                                            Data.BoundedSize.Class.IsBoundedSize 0
                                                                                                 4
                                                                                                 a_0) =>
                                              Data.BoundedSize.Class.BoundedSize 0 4 a_0

typeFromInt :: Int -> Type

Transform a data-level Int to a Type value LitT (NumTyLit 3)

>>> typeFromInt 3
LitT (NumTyLit 3)
>>> ppr $ typeFromInt 3
3