Data.BoundedSize

Description

sized-text combinators are defined for members of IsBoundedSize class. The package includes IsBoundedSize instances for several common types.

This module is meant to be imported qualifed, e.g.

import qualified Data.BoundedSize as S

Constructing bounded-size values

See also unsafeCreate

create :: (IsBoundedSize l u a, KnownNat l, KnownNat u) => a -> Maybe (BoundedSize l u a)

Working with bounded-size values

size :: IsBoundedSize l u a => a -> Int

IsBoundedSize class

data family BoundedSize l u a

Data family which wraps values of the underlying type giving them a type-level size. BoundedSize 6 10 t means a value of type t of size between 6 and 10.

class IsBoundedSize (l :: Nat) (u :: Nat) a where

Class of types which can be assigned a type-level minimum and maximum length.

Minimal complete definition

unsafeCreate, unwrap, size, predicate

Methods

unwrap :: BoundedSize l u a -> a

Forget type-level minimum and maximum size, obtaining the underlying value.

FixedSize class @FixedSize l a == BoundedSize l l a

type FixedSize (l :: Nat) a = BoundedSize l l a

Class of types which can be assigned a fixed type-level size.

class IsBoundedSize l u a => IsBoundedSizeText l u a where

Class of types which can be assigned a type-level minimum and maximum length.

Associated Types

type Elem a

Data family which wraps values of the underlying Text-like type giving them a type-level length. BoundedSizeText 6 10 t means a value of type t of length between 6 and 10.

Basic element type. For IsBoundedSizeTextText [a], this is a.

Methods

length :: a -> Int

append :: a -> a -> a

replicate :: Int -> Elem a -> a

map :: (Elem a -> Elem a) -> a -> a

take :: Int -> a -> a

drop :: Int -> a -> a

type FixedSizeText (l :: Nat) a = IsBoundedSizeText l l a => FixedSize l a