Data.BoundedSize

Description

sized-text combinators are defined for members of IsBoundedSize class. The package includes IsBoundedSize instances for several common types.

This module is meant to be imported qualifed, e.g.

import qualified Data.BoundedSize as S

Documentation

class HasSize a where

Size is an abstract Int-valued property of a Type that can be used to refine values of that type.

Methods

size :: a -> Int

class HasSize a => IsBytes a where

Minimal complete definition

append, replicate, map, take, drop

Associated Types

type Elem a

Basic element type. For IsBoundedSizeBytesText [a], this is a.

Methods

length :: a -> Int

append :: a -> a -> a

replicate :: Int -> Elem a -> a

map :: (Elem a -> Elem a) -> a -> a

take :: Int -> a -> a

drop :: Int -> a -> a

class (KnownNat l, KnownNat u, HasSize a) => IsBoundedSize l u a

Class of types which can be assigned a type-level minimum and maximum size.

Associated Types

data BoundedSize l u a

class IsBoundedSize l l a => IsFixedSize l a

Class of types which can be assigned a fixed type-level size.

type FixedSize n a = BoundedSize n n a

class IsBoundedSize 0 l a => IsMaxSize l a

Class of types which can be assigned a maximum type-level size.

type MaxSize n a = BoundedSize 0 n a

class (IsBoundedSize l u a, IsBytes a) => IsBoundedSizeBytes l u a

Class of types which can be assigned a type-level minimum and maximum length.

class (IsFixedSize n a, IsBytes a) => IsFixedSizeBytes n a

class (IsMaxSize n a, IsBytes a) => IsMaxSizeBytes n a

type AtLeast c m = forall l. (KnownNat l, m <= l) => c l

Sometimes a MaxSize value is given, but the only thing known about the "max" is that the value is legal, e.g. "foo" could be a MaxSize 3 String, or a MaxSize 4 String, so it is AtLeast MaxSize 3 Then it can be later cast to a specific MaxSize (of at least 3).

fx :: LitS -> Q Exp

Type-safe constructor for fixed-length string literals: BoundedSize a l l

mx :: LitS -> Q Exp

Type-safe constructor for bounded-length string literals: BoundedSize a 0 l

>>> :t "Foobar"
"Foobar" :: IsString p => p
>>> :t LitS "Foobar"
LitS "Foobar" :: LitS
>>> :t mx "Foobar"
mx "Foobar" :: Q Exp
>>> $(mx "Foobar") :: BoundedSize 0 6 ByteString
"Foobar"

>>> runQ $ ppr <$> mx "Foobar"
Data.Smart.unsafeCreate "Foobar" :: forall a_0 . (Data.HasSize.HasSize a_0,
                                                  Data.String.IsString a_0) =>
                                    Data.BoundedSize.Class.BoundedSize 0 6 a_0

>>> :t runQ $ ppr <$> mx "Foobar"
runQ $ ppr <$> mx "Foobar" :: Quasi m => m Doc

where 6 is the string length obtained at compile time.

unsafeCreateExp :: (Type -> Type -> Name -> Type) -> Type -> Type -> Name -> String -> Exp

Construct > unsafeCreate Foobar :: forall a. (IsString a, HasSize a) => typef a l u where l and u are the type-level KnownNat versions of the bounds of s >>> :set -Wno-name-shadowing >>> at <- runQ $ newName "a" >>> ppr $ unsafeCreateExp universallyQuantifiedBoundedSizeType (typeFromInt 0) (typeFromInt 4) at "Boo!" Data.Smart.unsafeCreate "Boo!" :: forall a_0 . (Data.HasSize.HasSize a_0, Data.String.IsString a_0) => Data.BoundedSize.Class.BoundedSize 0 4 a_0

typeFromInt :: Int -> Type

Transform a data-level Int to a Type value

>>> typeFromInt 3
LitT (NumTyLit 3)
>>> ppr $ typeFromInt 3
3