Data.BoundedSize

Description

The BoundedSize l u instance of Predicate and various specializations; sized-text combinators are defined for members of IsBoundedSize class. The package includes IsBoundedSize instances for several common types.

This module is meant to be imported qualifed, e.g.

import qualified Data.BoundedSize as S

Documentation

class HasSize a where

size is an abstract Int-valued property of a type that can be used to refine values of that type.

The size of an Integer is roughly the number of bytes required to represent

The size of a ListLike is just the number of Items

Methods

size :: a -> Int

newtype BoundedSize (l :: Nat) (u :: Nat) a

Class of types which can be assigned a type-level minimum and maximum size.

Constructors

BoundedSize a

class (KnownNat l, KnownNat u, HasSize a) => IsBoundedSize l u a

class IsBoundedSize l l a => IsFixedSize l a

Class of types which can be assigned a fixed type-level size.

type FixedSize n = BoundedSize n n

class IsBoundedSize 0 l a => IsMaxSize l a

Class of types which can be assigned a maximum type-level size.

type MaxSize n = BoundedSize 0 n

type AtLeast c m = forall l. (KnownNat l, m <= l) => c l

Sometimes a MaxSize value is given, but the only thing known about the "max" is that the value is legal, e.g. "foo" could be a MaxSize 3 String, or a MaxSize 4 String, so it is AtLeast MaxSize 3 Then it can be later cast to a specific MaxSize (of at least 3).

fx :: LitS -> Q Exp

Type-safe constructor for fixed-length string literals: a ? BoundedSize l l

mx :: LitS -> Q Exp

Type-safe constructor for bounded-length string literals: a ? BoundedSize 0 l

>>> :t "Foobar"
"Foobar" :: IsString p => p
>>> :t LitS "Foobar"
LitS "Foobar" :: LitS
>>> :t mx "Foobar"
mx "Foobar" :: Q Exp
>>> $(mx "Foobar") :: ByteString ? BoundedSize 0 6
"Foobar"

>>> runQ $ ppr <$> mx "Foobar"
Data.Refined.unsafeCreate "Foobar" :: forall a_0 . (Data.HasSize.HasSize a_0,
                                                    Data.String.IsString a_0) =>
                                      (Data.Refined.?) a_0 (Data.BoundedSize.Class.BoundedSize 0 6)

>>> :t runQ $ ppr <$> mx "Foobar"
runQ $ ppr <$> mx "Foobar" :: Quasi m => m Doc

where 6 is the string length obtained at compile time.

unsafeCreateExp :: (Type -> Type -> Name -> Type) -> Type -> Type -> Name -> String -> Exp

Construct unsafeCreate Foobar :: forall a. (IsString a, HasSize a) => typef l u a where l and u are the type-level KnownNat versions of the bounds of s

>>> :set -Wno-name-shadowing
>>> at <- runQ $ newName "a"
>>> ppr $ unsafeCreateExp universallyQuantifiedBoundedSizeType (typeFromInt 0) (typeFromInt 4) at "Boo!"
Data.Refined.unsafeCreate "Boo!" :: forall a_0 . (Data.HasSize.HasSize a_0,
                                                  Data.String.IsString a_0) =>
                                    (Data.Refined.?) a_0 (Data.BoundedSize.Class.BoundedSize 0 4)

typeFromInt :: Int -> Type

Transform a data-level Int to a Type value

>>> typeFromInt 3
LitT (NumTyLit 3)
>>> ppr $ typeFromInt 3
3

class (IsBoundedSize l u a, ListOps a) => IsBoundedSizeListOps l u a

Class of types which can be assigned a type-level minimum and maximum length.

class (IsFixedSize n a, ListOps a) => IsFixedSizeListOps n a

class (IsMaxSize n a, ListOps a) => IsMaxSizeListOps n a

append :: forall l1 u1 l2 u2 a. (IsBoundedSizeListOps l1 u1 a, IsBoundedSizeListOps l2 u2 a, IsBoundedSizeListOps (l1 + l2) (u1 + u2) a) => (a ? BoundedSize l1 u1) -> (a ? BoundedSize l2 u2) -> a ? BoundedSize (l1 + l2) (u1 + u2)

Append two BoundedSize values together.

>>> append $(mx "foo") $(fx "bear")
"foobear"

>>> :type append $(mx "foo") $(fx "bear")
append $(mx "foo") $(fx "bear")
  :: (Data.ListLike.Base.ListLike a (Item a), HasSize a,
      Data.String.IsString a) =>
     a ? BoundedSize 4 7

>>> append $(fx "Hello, ") $(mx "world!")
"Hello, world!"

replicate :: forall l u a. IsBoundedSizeListOps l u a => Item a -> a ? BoundedSize l u

Construct a new BoundedSize of maximum length from a basic element.

>>> replicate '=' :: String ? BoundedSize 5 10
"=========="

map :: forall l u a. IsBoundedSizeListOps l u a => (Item a -> Item a) -> (a ? BoundedSize l u) -> a ? BoundedSize l u

Map a BoundedSize to a BoundedSize of the same length.

>>> map toUpper $(fx "Hello") :: String ? BoundedSize 5 5
"HELLO"

take :: forall l1 u1 l2 u2 a. (IsBoundedSizeListOps l1 u1 a, IsBoundedSizeListOps l2 u2 a) => (a ? BoundedSize l1 u1) -> a ? BoundedSize l2 u2

Reduce BoundedSize length, preferring elements on the left.

>>> take $(mx "Foobar") :: String ? BoundedSize 0 3
"Foo"
>>> :t take $(fx "Hello")
take $(fx "Hello")
  :: (Data.ListLike.Base.ListLike a (Item a), KnownNat l2,
      KnownNat u2, HasSize a, Data.String.IsString a) =>
     a ? BoundedSize l2 u2

>>> :t take $(fx "Hello") :: ByteString ? BoundedSize 4 32
take $(fx "Hello") :: ByteString ? BoundedSize 4 32
  :: ByteString ? BoundedSize 4 32
>>> take $(fx "Hello") :: ByteString ? BoundedSize 4 32
"Hello"
>>> length (take $(fx "Hello") :: ByteString ? BoundedSize 4 32)
5
>>> take $(fx "Hello") :: ByteString ? BoundedSize 4 32
"Hello"
>>> take $(fx "Hello") :: ByteString ? BoundedSize 4 4
"Hell"
>>> take $(fx "HelloHelloHelloHelloHelloHelloHello") :: ByteString ? BoundedSize 4 32
"HelloHelloHelloHelloHelloHelloHe"

drop :: forall l1 u1 l2 u2 a. (IsBoundedSizeListOps l1 u1 a, IsBoundedSizeListOps l2 u2 a) => (a ? BoundedSize l1 u1) -> a ? BoundedSize l2 u2

Reduce BoundedSize length, preferring elements on the right.

>>> drop $(fx "Foobar") :: String ? BoundedSize 2 2
"ar"
>>> drop $(mx "Foobar") :: String ? BoundedSize 0 2
"ar"

bounds :: forall l u a. IsBoundedSize l u a => (a ? BoundedSize l u) -> (Int, Int)

Obtain length bounds from the type.

length :: forall l u a. IsBoundedSizeListOps l u a => (a ? BoundedSize l u) -> Int

Obtain value-level length. Consult the plain data value.

padLeft :: forall l1 u1 l2 u2 a. (IsBoundedSizeListOps l1 u1 a, IsBoundedSizeListOps l2 u2 a, IsBoundedSizeListOps 0 (u2 + u1) a) => Item a -> (a ? BoundedSize l1 u1) -> a ? BoundedSize l2 u2

Fill a BoundedSize with extra elements up to target length, padding original elements to the left.

padRight :: forall l1 u1 l2 u2 a. (IsBoundedSizeListOps l1 u1 a, IsBoundedSizeListOps l2 u2 a, IsBoundedSizeListOps 0 (u1 + u2) a) => Item a -> (a ? BoundedSize l1 u1) -> a ? BoundedSize l2 u2

Like padLeft, but original elements are padded to the right.

data Refined p a

plain :: IsRefined p a => Refined p a -> a

unsafeCreate :: IsRefined p a => a -> Refined p a

type (?) a p = Refined p a infixr 1

An infix alias for Refined.

class Coercible a (p a) => Predicate (p :: Type -> Type) a where

Minimal complete definition

predicate, failMsg

Methods

predicate :: p a -> Bool

failMsg :: p a -> String

safeCreate :: IsRefined p a => a -> Either String (Refined p a)

create :: IsRefined p a => a -> Maybe (Refined p a)

refine :: IsRefined p a => a -> Refined p a

pred :: a -> Bool

fromNat :: (Num b, KnownNat n) => proxy n -> b

Type-level KnownNat to value-level Num

>>> fromNat (Proxy @5) :: Int
5