Data.BoundedSize.Class

Description

Use this module when you need to add an IsBoundedSize instance to a type.

Documentation

class HasSize a where

Size is an abstract Int-valued property of a Type that can be used to refine values of that type.

Methods

size :: a -> Int

class HasSize a => IsBytes a where

Minimal complete definition

append, replicate, map, take, drop

Associated Types

type Elem a

Basic element type. For IsBoundedSizeBytesText [a], this is a.

Methods

length :: a -> Int

append :: a -> a -> a

replicate :: Int -> Elem a -> a

map :: (Elem a -> Elem a) -> a -> a

take :: Int -> a -> a

drop :: Int -> a -> a

class HasSize a => IsBoundedSize (l :: Nat) (u :: Nat) a

Class of types which can be assigned a type-level minimum and maximum length.

Associated Types

data BoundedSize l u a

class IsBoundedSize l l a => IsFixedSize l a

Class of types which can be assigned a fixed type-level size.

type FixedSize n a = BoundedSize n n a

class IsBoundedSize 0 l a => IsMaxSize l a

Class of types which can be assigned a maximum type-level size.

type MaxSize n a = BoundedSize 0 n a

class (IsBytes a, IsBoundedSize l u a) => IsBoundedSizeBytes l u a

Class of types which can be assigned a type-level minimum and maximum length.

class IsBoundedSizeBytes l l a => IsFixedSizeBytes l a

class IsBoundedSizeBytes 0 l a => IsMaxSizeBytes l a