Data.BoundedSize.Class

Description

Use this module when you need to add an IsBoundedSize instance to a type.

Documentation

class IsBytes a where

Class of types with ByteString-like operations

Associated Types

type Elem a

Methods

length :: a -> Int

append :: a -> a -> a

replicate :: Int -> Elem a -> a

map :: (Elem a -> Elem a) -> a -> a

take :: Int -> a -> a

drop :: Int -> a -> a

class IsBoundedSize (l :: Nat) (u :: Nat) a where

Class of types which can be assigned a type-level minimum and maximum length.

Minimal complete definition

unsafeCreate, unwrap, size, predicate

Associated Types

data BoundedSize l u a

Data family which wraps values of the underlying type giving them a type-level size. BoundedSize t 6 10 means a value of type t of size between 6 and 10.

Methods

unsafeCreate :: a -> BoundedSize l u a

Simply wrap a value in a BoundedSize as is, assuming any length.

WARNING Use it only when you know what you're doing.

For example, an expression like

unsafeCreate "somestring" :: BoundedSize String 50 100

will typecheck, although the stored size information will not match actual value size. This may result in wrong behaviour of all functions defined for IsBoundedSize.

When writing new IsBoundedSize instances, make this simply apply the constructor of BoundedSize.

unwrap :: BoundedSize l u a -> a

Forget type-level minimum and maximum size, obtaining the underlying value.

size :: a -> Int

predicate :: (KnownNat l, KnownNat u) => BoundedSize l u a -> Either String (BoundedSize l u a)

safeCreate :: (KnownNat l, KnownNat u) => a -> Either String (BoundedSize l u a)

create :: (KnownNat l, KnownNat u) => a -> Maybe (BoundedSize l u a)

type FixedSize n a = BoundedSize n n a

Class of types which can be assigned a fixed type-level size.

class (IsBoundedSize l u a, IsBytes a) => IsBoundedSizeBytes l u a

Class of ByteString-like types which can be assigned a type-level minimum and maximum length.

type FixedSizeBytes n a = IsBoundedSizeBytes n n a => FixedSize n a