Data.BoundedSize.Class

Description

Use this module when you need to add an IsBoundedSize instance to a type.

Documentation

class HasSize a where

size is an abstract Int-valued property of a type that can be used to refine values of that type.

The size of an Integer is roughly the number of bytes required to represent

The size of a ListLike is just the number of Items

Methods

size :: a -> Int

newtype BoundedSize (l :: Nat) (u :: Nat) a

Class of types which can be assigned a type-level minimum and maximum size.

Constructors

BoundedSize a

class (KnownNat l, KnownNat u, HasSize a) => IsBoundedSize l u a

class IsBoundedSize l l a => IsFixedSize l a

Class of types which can be assigned a fixed type-level size.

type FixedSize n = BoundedSize n n

class IsBoundedSize 0 l a => IsMaxSize l a

Class of types which can be assigned a maximum type-level size.

type MaxSize n = BoundedSize 0 n

type AtLeast c m = forall l. (KnownNat l, m <= l) => c l

Sometimes a MaxSize value is given, but the only thing known about the "max" is that the value is legal, e.g. "foo" could be a MaxSize 3 String, or a MaxSize 4 String, so it is AtLeast MaxSize 3 Then it can be later cast to a specific MaxSize (of at least 3).

data Refined p a

unsafeCreate :: IsRefined p a => a -> Refined p a

plain :: IsRefined p a => Refined p a -> a

type (?) a p = Refined p a infixr 1

An infix alias for Refined.

class Coercible a (p a) => Predicate (p :: Type -> Type) a where

Minimal complete definition

predicate, failMsg

Methods

predicate :: p a -> Bool

failMsg :: p a -> String

safeCreate :: IsRefined p a => a -> Either String (Refined p a)

create :: IsRefined p a => a -> Maybe (Refined p a)

refine :: IsRefined p a => a -> Refined p a

pred :: a -> Bool

fromNat :: (Num b, KnownNat n) => proxy n -> b

Type-level KnownNat to value-level Num

>>> fromNat (Proxy @5) :: Int
5