Data.SizedText.Class

Description

Use this module when you need to add an IsSizedText instance to a type.

Documentation

class IsSizedText a where

Class of types which can be assigned a type-level minimum and maximum length.

Associated Types

data Sized a (l :: Nat) (u :: Nat)

Data family which wraps values of the underlying type giving them a type-level length. Sized t 6 10 means a value of type t of length between 6 and 10.

type Elem a

Basic element type. For IsSizedText [a], this is a.

Methods

unsafeCreate :: a -> Sized a l u

Simply wrap a value in a Sized as is, assuming any length.

WARNING Use it only when you know what you're doing.

For example, an expression like

unsafeCreate "somestring" :: Sized String 50 100

will typecheck, although the stored length information will not match actual string size. This may result in wrong behaviour of all functions defined for IsSizedText.

When writing new IsSizedText instances, make this simply apply the constructor of Sized.

unwrap :: Sized a l u -> a

Forget type-level minimum and maximum length, obtaining the underlying value.

length :: a -> Int

append :: a -> a -> a

replicate :: Int -> Elem a -> a

map :: (Elem a -> Elem a) -> a -> a

take :: Int -> a -> a

drop :: Int -> a -> a

type Static a (l :: Nat) = Sized a l l

Class of types which can be assigned a type-level fixed length.

createLeft :: forall a l u. (IsSizedText a, KnownNat l, KnownNat u) => Elem a -> a -> Sized a l u

Elements on the left are preferred. >>> createLeft ' ' "foobarbaz" :: Sized String 0 6 "foobar" >>> createLeft '@' "foobarbaz" :: Sized String 12 20 "foobarbaz@@@"

createRight :: forall a l u. (IsSizedText a, KnownNat l, KnownNat u) => Elem a -> a -> Sized a l u

Just like createLeft, except that elements on the right are preferred. >>> createRight '@' "foobarbaz" :: Sized String 0 6 "barbaz" >>> createRight # "foobarbaz" :: Sized String 12 20 "###foobarbaz"

create :: forall a (l :: Nat) (u :: Nat). (IsSizedText a, KnownNat l, KnownNat u) => a -> Maybe (Sized a l u)

Attempt to safely create a Sized if it matches target length.

>>> create "foobar" :: Maybe (Sized String 6 10)
Just "foobar"
>>> create "barbaz" :: Maybe (Sized String 0 4)
Nothing

This is safer than unsafeCreate and unlike with createLeft createRight the source value is left unchanged. However, this implies a further run-time check for Nothing values.