System.SED.MCTP.Common.TypeUIDs.TH

Description

Template Haskell for parsing Table column types in Section 5.1.3.

Documentation

ttype :: QuasiQuoter

QuasiQuoter for Type Table row tables.

Each row of the Type Table is in its own table in the Core Spec.

ttypeDecs :: String -> [Dec]

typeTableParser :: Parser TypeTableRow

typeTableRow :: [Int] -> Parser TypeTableRow

dTypeTableRow :: TypeTableRow -> [Dec]

typeTableTitle :: Parser ByteString

spaces :: Parser ByteString

rowSepFieldLengths :: Parser [Int]

rowSep :: Parser [Int]

blankLines :: Parser ()

tableRowFields :: [Int] -> Parser [ByteString]

data TypeTableRow

Constructors

TypeTableRow TypeUIDField TypeName FormatString

type TypeUIDField = ByteString

type TypeName = ByteString

type FormatString = ByteString

header :: [Int] -> Parser ()

type Type_UID = UID

type Base_Type_UID = Type_UID

type Non_Base_Type_UID = Type_UID

type Named_Value_Type_UID = Type_UID

type Core_type_def_max_enum_ranges = 16

type Core_type_def_max_alternative_uidrefs = 16

type Core_type_def_max_byte_table_uidrefs = 16

type Core_type_def_max_object_table_uidrefs = 16

type Core_type_def_max_struct_uidrefs = 16

type Core_type_def_max_set_ranges = 16

data Core_Format (n :: Nat) :: Type where

Constructors

Base_Type_Format :: Core_Format 0

Simple_Type_Format :: Base_Type_UID -> Core_uinteger_2 -> Core_Format 1

Enumeration_Type_Format :: ([(Core_uinteger_2, Core_uinteger_2)] ? BoundedSize 1 Core_type_def_max_enum_ranges) -> Core_Format 2

Alternative_Type_Format :: (Set Non_Base_Type_UID ? BoundedSize 2 Core_type_def_max_alternative_uidrefs) -> Core_Format 3

List_Type_Format :: Core_uinteger_2 -> Non_Base_Type_UID -> Core_Format 4

Restricted_Reference_Type_5_Format :: (Set Byte_Table_UID ? BoundedSize 1 Core_type_def_max_byte_table_uidrefs) -> Core_Format 5

Restricted_Reference_Type_6_Format :: (Set Object_Table_UID ? BoundedSize 1 Core_type_def_max_object_table_uidrefs) -> Core_Format 6

General_Reference_Type_7_Format :: Core_Format 7

General_Reference_Type_8_Format :: Core_Format 8

General_Reference_Type_9_Format :: Core_Format 9

General_Reference_Table_Type_Format :: Table_Kind -> Core_Format 10

Named_Value_Name_Type_Format :: Core_max_bytes_32 -> Non_Base_Type_UID -> Core_Format 11

Named_Value_Integer_Type_Format :: Core_integer_2 -> Non_Base_Type_UID -> Core_Format 12

Named_Value_Uinteger_Type_Format :: Core_uinteger_2 -> Non_Base_Type_UID -> Core_Format 13

Struct_Type_Format :: ([Named_Value_Type_UID] ? BoundedSize 1 Core_type_def_max_struct_uidrefs) -> Core_Format 14

Set_Type_Format :: ([(Core_uinteger_2, Core_uinteger_2)] ? BoundedSize 1 Core_type_def_max_set_ranges) -> Core_Format 15

data Some_Core_Format

Constructors

forall n.(Show (Core_Format n), KnownNat n) => Some_Core_Format (Core_Format n)

class Is_Core_Format a where

Methods

formatCode :: a -> Word8

type Type_Name = String

type Base_Type_Name = Type_Name

type Non_Base_Type_Name = Type_Name

type Named_Value_Type_Name = Type_Name

type Table_Name = String

type Byte_Table_Name = Table_Name

type Object_Table_Name = Table_Name

newtype One_Of s a

Constructors

One_Of a

class IsOne_Of s a where

Minimal complete definition

Nothing

Methods

refiningSet :: s a -> Set a

core_Base_Type_Names :: Set String

type Core_Base_Type_Names = Type

core_Table_Kind_Names :: Set String

type Core_Table_Kind_Names = Type

data Core_Format_Spec

Constructors

Base_Type_Spec (String ? One_Of Core_Base_Type_Names)

Simple_Type_Spec (Base_Type_Name, Core_uinteger_2)

Enumeration_Type_Spec ([(Core_uinteger_2, Core_uinteger_2)] ? BoundedSize 1 Core_type_def_max_enum_ranges)

Alternative_Type_Spec (Set Non_Base_Type_Name ? BoundedSize 2 Core_type_def_max_alternative_uidrefs)

List_Type_Spec (Core_uinteger_2, Non_Base_Type_Name)

Restricted_Reference_Type_5_Spec (Set Byte_Table_Name ? BoundedSize 1 Core_type_def_max_byte_table_uidrefs)

Restricted_Reference_Type_6_Spec (Set Object_Table_Name ? BoundedSize 1 Core_type_def_max_object_table_uidrefs)

General_Reference_Type_7_Spec

General_Reference_Type_8_Spec

General_Reference_Type_9_Spec

General_Reference_Table_Type_Spec (String ? One_Of Core_Table_Kind_Names)

Named_Value_Name_Type_Spec (Core_max_bytes_32, Non_Base_Type_Name)

Named_Value_Integer_Type_Spec (Core_integer_2, Non_Base_Type_Name)

Named_Value_Uinteger_Type_Spec (Core_uinteger_2, Non_Base_Type_Name)

Struct_Type_Spec ([Named_Value_Type_Name] ? BoundedSize 1 Core_type_def_max_struct_uidrefs)

Set_Type_Spec ([(Core_uinteger_2, Core_uinteger_2)] ? BoundedSize 1 Core_type_def_max_set_ranges)

stringSetParser :: (IsList l, Item l ~ String) => l -> Parser ByteString

formatSpecParser :: Parser Core_Format_Spec

Parse the string in the Format colum

baseTypeFormatSpecParser :: Parser Core_Format_Spec

simpleTypeFormatSpecParser :: Parser Core_Format_Spec

tenum :: QuasiQuoter

QuasiQuoter for Type Table enum tables.

Some enumerations are explicated by tables following the Type Tables that refer to them.

tenumDecs :: String -> [Dec]

dEnum :: (String, [EnumRow]) -> [Dec]

enumTableParser :: Parser (String, [EnumRow])

data EnumRow

Constructors

EnumRow String [Int]

enumTableTitle :: Parser String

enumTableRow :: [Int] -> Parser EnumRow

type U1 = "abc"

type U2 = "u2"

data Core_type (uid :: Symbol) a where

Constructors

Core_type_uinteger :: Core_type "00000005" (Core_uinteger_2 -> Implementation_uinteger)